requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
black==23.12.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dateparser==1.2.0
exceptiongroup==1.2.0
frozenlist==1.4.1
idna==3.6
iniconfig==2.0.0
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.3
packaging==23.2
pandas==2.1.4
pathspec==0.12.1
platformdirs==4.1.0
pluggy==1.3.0
polars==0.20.4
psycopg2-binary==2.9.9
pycryptodome==3.20.0
pytest==7.4.4
python-binance==1.0.19
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
regex==2023.12.25
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.25
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
ujson==5.9.0
urllib3==2.1.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
